PHPCS: fix up the code base [4] - multi-line function calls
Multi-line function calls need to have each argument on a new line.
In a next iteration of this principle, it is expected that a sniff will be introduced to ban multi-line function call arguments.

With this mind, a number of function calls with multi-line parameters which are currently already causing errors to be thrown by PHPCS, have been fixed by moving multi-line function call arguments out of the function call and defining these as a variable before passing it to the function call.