Fix specific license in a plugin
The doc says: 
```
If a plugin or theme specifies a license in their main plugin file or stylesheet, the comment looks like
this: 
Copyright (C) 2018 Example Plugin Author
This file is distributed under the GPLv2.
```
but it dosen't seems to work for plugins. It looks like the "License" param in get_file_headers was missing.